Add support for Python 3.12
Python 3.12 has been released on October 2023.
This commit installs it in the image, but still keeps Python 3.11
as the default for now.

Note that due to the removal of the long-deprecated `pkgutil.ImpImporter`,
doing `pip install --upgrade` with Python 3.12 fails with:

```
module 'pkgutil' has no attribute 'ImpImporter'. Did you mean: 'zipimporter'?
```

To avoid this, we fallback to `ensurepip` if `pip` fails to upgrade itself.
See https://docs.python.org/3/whatsnew/3.12.html#importlib