Make executables installed with pip available
Recent Debian/Ubuntu versions have modified pip to use the "User Scheme" by default.
This commit thus adds `/home/tox/.local/bin` to the PATH, so that say "pytest" is available
as a command after `pip install pytest`.

Without this, pytest is installed but not available as a standalone command.